requirements.txt
blinker==1.9.0
click==8.3.0
Flask==3.1.2
Flask-Login==0.6.3
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.2
MarkupSafe==3.0.3
numpy==2.2.6
opencv-python==4.12.0.88
ptext==1.1.0
pypdf==6.1.3
python-docx==1.2.0
typing_extensions==4.15.0
Werkzeug==3.1.3
